ex_9/kajiwara/requirements.txt
numpy==1.20.3
pandas==1.2.4
librosa==0.8.1
SoundFile==0.10.3.post1
torchinfo==0.1.4
tqdm==4.61.0
matplotlib==3.4.2
tensorboard==2.5.0
hydra-core==1.1.0
conformer==0.2.5
